Move dum class to tests folder
